twitter/data/pipeline/requirements.txt
vaex>=4.16.0
python_bidi>=0.4.2
pandas>=1.5.3
numpy>=1.24.3
arabic_reshaper>=3.0.0
pyspark>=3.4.0
dask>=2023.3.2
matplotlib>=3.7.1
-e git+ssh://git@github.com/ilex-paraguariensis/yerbamate.git@b53e93b46537e7de75c8c8cd42f142cfabbaf96b#egg=yerbamate
ipdb>=0.13.13
